Moved outdated matrix functions to tests (mask_gt, mask_to_ignore and so on)
Moved outdated matrix functions to tests (mask_gt, mask_to_ignore and so on)
